users module dir refactor
