+++ src/main/java/com/trymad/litechess_monolith/chessgame/internal/client/UserInfoClient.java
@@ -2,7 +2,7 @@ package com.trymad.litechess_monolith.chessgame.internal.client;
 
 import java.util.UUID;
 
-import com.trymad.litechess_monolith.users.UserInfoDTO;
+import com.trymad.litechess_monolith.users.api.dto.UserInfoDTO;
 
 public interface UserInfoClient {
