+++ src/main/java/com/trymad/litechess_monolith/chessgame/internal/service/ChessPartyService.java
@@ -21,7 +21,7 @@ import com.trymad.litechess_monolith.chessgame.internal.model.LiveGame;
 import com.trymad.litechess_monolith.chessgame.internal.repository.ChessPartyRepository;
 import com.trymad.litechess_monolith.matchmaking.api.event.GameFindedEvent;
 import com.trymad.litechess_monolith.shared.event.EventPublisher;
-import com.trymad.litechess_monolith.users.UserInfoDTO;
+import com.trymad.litechess_monolith.users.api.dto.UserInfoDTO;
 import com.trymad.litechess_monolith.websocket.api.event.MoveEvent;
 
 import lombok.RequiredArgsConstructor;
