+++ src/main/java/com/trymad/litechess_monolith/users/api/dto/UserInfoCreateDTO.java
@@ -1,4 +1,4 @@
-package com.trymad.litechess_monolith.users.internal;
+package com.trymad.litechess_monolith.users.api.dto;
 
 public record UserInfoCreateDTO(String nickname) {
