+++ src/main/java/com/trymad/litechess_monolith/users/api/package-info.java
@@ -1,0 +1,7 @@
+@org.springframework.modulith.ApplicationModule(
+  type = Type.OPEN
+)
+package com.trymad.litechess_monolith.users.api;
+
+import org.springframework.modulith.ApplicationModule.Type;
+
